requirements.txt
docopt-ng
mne
mne-bids
bids-validator
